Add recovery from NFS ingest failure
Summary:
I still observe at time that the edits.new cannot be seen by the
standby. This potentially creates a regression from the federation
branch.

In this patch, after upload of the checkpoint is done, and we see that
there is no ingest, we can instantiate ingest for edits -> since after
rolling the image edits.new was turned into edits. This basically will
mean that if we can instantiate edits.new after saving namespace, it's
going to help. If we fail there, we will do this after uploading. By
doing so, in the worst case we fallback to the federation behaviour.

-added check for instantiateIngest, in which if the file does not exist,
we don't change the state -> otherwise standby sets the state to
ingesting edits.new, ingest is not null, and standby thinks that ingest
is working, where in fact
it's not, and this is not a correct behavior -> ingest won't be
recreated until next checkpoint !!!

-ported retries for edits.new to trunk, since they were only committed
to alligator, increased the number of retries to 30

-added test case to simulate edits.new failure

-added some logging for failover, fixed some comment issues

Test Plan: all avatar tests

Reviewers: hkuang

Reviewed By: hkuang

Task ID: 1039092