recoverLease should return true if the file has no blocks
Summary: This makes recoverLease return true if the file is found to have no blocks and is finalized as part of the recoverLease call. Previously recoverLease would return false and the client would have to retry in such case.

Test Plan: Run on dev cluster with HBase workload

Reviewers: tomasz, hkuang

Reviewed By: hkuang

Task ID: 774638