Fix deadlock in node manager.
Summary:
We should not try to acquire a node lock while holding its node
container lock. This is because there are other code paths where the
node lock is taken first and then the node container lock.

Fix the deadlock by having a separate data structure to hold the
runnable nodes of a type. This removes the need for iterating over the
node containers.

Test Plan:
test cluster

Revert Plan:

Reviewers: dms, aching, pyang

Reviewed By: dms

Task ID: 1126283