Remove interrupting and joining the emptier
Summary:
We need to remove the interruption and joining of the emptier, it's
using the client, and we can't be sure if the interruption would not be
swallowed anywhere, and I have seen the failover hanging on emptier
join().

-we are cleanly shutting down rpc threads so this should be enough, and
the emptier thread is a daemon, which will be exited anyways on the JVM
exit.

Test Plan: deploy

Reviewers: hkuang, pritam

Reviewed By: hkuang