Use the correct history directory.
Summary:
When the remote job tracker starts, it gets a new session ID, and a
corresponding session log path. But it should use the session log path
for the parent job tracker.

Normally this does not make a difference since the session IDs of the
parent and remote job trackers fall in the same 1000 (number of sessions
per directory). But when they fall in different buckets, users will not
be able to see the job history page of finished jobs

Revert Plan:

Reviewers: dms, aching, pyang

Reviewed By: dms