Timeout local RPC server in federated mode.
Summary:
Provide a configurable way (off by default) to timeout the RPC server of
the local job tracker when there is a remote job tracker running.
The timeout of the RPC server is not done normally, since the client
process is expected to exit after the job is done.
But for long running clients like RAID, we need to actually close the
RPC server after some time.
This behavior is controlled through configuration.

Test Plan: tested on test cluster raidnode

Reviewers: dms, aching, pyang

Reviewed By: dms

CC: weiyan