Disallow DataNode to format itself when data.dir is not empty
Summary:
When a data dir is not empty, the data node should exit rather than
automatically reformat the data dir. Otherwise leads to inconsistent
behaviour and potential data loss.

Test Plan:
Set data.dir in hdfs-site.xml to a non-empty directory, containing
data that does not belong to the DataNode. Attempt to start DataNode
and verify that an error message is logged and DataNode shuts down.

Revert Plan:

Reviewers: hkuang

Reviewed By: hkuang

Task ID: 1139555