Add a raidshell command to raid a list of paths
Summary:
Add a command in this format:
-distRaid <raid_policy_name> <path1> ... <pathn>

that submit a map/reduce job to raid the given list of paths. A path could be a file or a directory. if it is a directory, only its children files are going to be raided.

Test Plan: Add a unit test in TestRaidShell

Reviewers: weiyan, dikang

Reviewed By: weiyan