HDFS: recounting INode count and disk space before exiting safe mode
Summary:
Standby Namenode's disk space quota is not accurate, as it misses addAdditionalBlock operation compared to primary. So the number needs to be recount before it becomes primary.

One thing I'm not sure is the cost of the method. Shall we complicate the codes a little bit to avoid one unnecessary recounting? Or we can do it as a follow-up.

Test Plan: Deploy in cluster.

Reviewers: hkuang, tomasz

Reviewed By: hkuang

Task ID: 1127167