Prevent CM -> Job Tracker timeouts.
Summary:
The cluster manager has a timeout on getting a response from the job
tracker on thrift calls. If the job tracker is doing something time
intensive and is not able to complete the thrift call, it can be timed
out.

This patch tries to prevent this situation from occuring. We add a
SessionDriverService.Iface implementation that simply queues up the
calls from the Cluster Manager during the thrift call. This will
guarantee quick response times to the cluster manager. There is a
separate thread to take the calls from the queue and execute them
in-order.

Keeping the calls in the same queue is important since that will reflect
the sequence of calls that the Cluster Manager makes.

Test Plan:
unit tests

Revert Plan:

Reviewers: dms, aching, pyang

Reviewed By: dms