Fix bug in detecting dead job tracker.
Summary:
The job tracker reporter thread in the corona task tracker should
periodically attempt a heartbeat to the corona job tracker to detect
a dead job tracker. The job tracker reporter has logic to heartbeat
every 3 minutes even if there are no tasks running for that job, but
there is a bug in that logic which is fixed by this diff.

Also:
1. Set a name for the job tracker reporter thread. This will help in
   debugging.
1. prefix log messages from job tracker reporter with the reporter name.

Test Plan:
unit tests

Revert Plan:

Reviewers: dms, aching, pyang

Reviewed By: dms