Remove synchronization from getListenerAddress.
Summary:
The primary avatar had a bug where it would enter a deadlock
since Server.waitForHandlers() is synchronized and would take the lock
and wait for all handlers and then a RPC handler could call
getListenerAddress() and wait for the lock on the Server object. Looks
like getListenerAddress() should not be synchronized since it does not
modify any state.

Test Plan: 1) All unit tests.

Reviewers: hkuang, weiyan, tomasz

Reviewed By: hkuang

Task ID: 1155907