Option to force failover for Avatar.
Summary:
Sometimes the primary Avatar does not shutdown cleanly and we
need an option to force a failover without any transaction id checking.
This can be done under a case where an operator can inspect the edit
files on the filer and the local disk and then force a failover without
any checking.

Test Plan:
1) Unit tests added.
2) All avatar unit tests.

Revert Plan:

Reviewers: hkuang, tomasz, weiyan

Reviewed By: hkuang