Putting ClusterManager into Safe Mode
Summary:
For now, we are able to set and unset the Safe Mode. We can fire the command to coronaadmin to set the safe mode, which in turn, asks the ClusterManager to go
into Safe Mode. The ClusterManager throws SafeModeExceptions on any new requests while it is in Safe Mode, and communicates to the ProxyJobTracker that it is going into
Safe Mode.

CoronaJobTrackers upon contacting the ClusterManager come to know that it is in Safe Mode. If they are not able to connect to the ClusterManager, while it is
down for an upgrade, they try the ProxyJobTracker to find if the ClusterManager has gone down for an upgrade. While it is up, the ClusterManager does not time out any of
the sessions and the nodes.

Test Plan:
Turn on the Safe Mode by:
hadoop coronaadmin -setSafeMode

Turn off the Safe Mode by:
hadoop coronaadmin -unsetSafeMode

1. If you turn on the Safe Mode while a job is running, it will pause and wait for the CM to come back up in the normal state. If you turn off the Safe Mode without
bringing down the ClusterManager, the execution will resume normally. Otherwise, currently the job fails because an InvalidSessionHandle exception in thrown, which will be
resolved once we persist the state of the ClusterManager and restore it when it comes back up.

2. If a job is started after the Safe Mode is turned on, it will wait for the CM to come back up in the normal state.

Reviewers: dms, aching, rvadali

Reviewed By: dms

Task ID: 1112019