[Raid Dir-Raid] Web UI changes for directory raid
Summary:
For directory raid, we may be more interested in the number of RAIDED,
NOT_RAIDED_TOO_NEW, NOT_RAIDED_TOO_SMALL, NOT_RAIDED_BUT_SHOULD directories
than that of files.

Change the webui to adopt the codec design and counters to include the counter of
directory.

The counters will be updated by the StatisticsCollector, which is another task.

Test Plan: deploy

Reviewers: dikang, hkuang

Reviewed By: hkuang

Task ID: 1078275