[Raid Dir-Raid] PlacementMonitor changes
Summary:
When PurgeMonitor scans the parity files, it will get the source file/directory
for each parity file and then compute the histogram of the blocks for them.
Change placementMonitor to get the block information for directory and
refactor BlockAndDatanodeResolver to work with multiple files.

Test Plan: Add a unit test testDirectoryPlacementMonitor

Reviewers: dikang, hkuang

Reviewed By: hkuang

Task ID: 1078264