[raid] the DistributedRaidFileSystem should deal with multi codec in one directory and the har directories.
Summary:
If a directory contains the files raided using both XOR and RS, we
should move all the parity files together with the source directory,
like the delete and undelete methods.

For the har one, we will throw the IOException.

Test Plan:
ant test, run locally on dev server.

will come up a unit test for the har one.

Reviewers: weiyan, hkuang

Reviewed By: weiyan