Map tasks do not need to connect to corona JT directly.
Summary:
The purpose of DirectTaskUmbilicalProtocol is to enable reducers to get
task completion events directly from the corona job tracker.
The mappers do not need to go through this. If they do, a large number
of concurrently mappers would start connecting to the corona JT and they
could get connection errors because that would overwhelm the corona JT

Test Plan: deploy

Reviewers: gauravmenghani, aching, pyang, dms

Task ID: 1166249