Fix TestAvatarFailover timeouts.
Summary:
TestAvatarFailover had an issue in double failovers where when
we restart the standby it might backoff full block reports and due to a
large block report delay, the blocks wouldn't be reported in time and
the standby would not be able to failover. Reducing the full block
report delay for the double failvoer tests.

Test Plan: Run it

Reviewers: tomasz, hkuang

Reviewed By: hkuang