[HDFS Tool] Add -u option in dfs -touch to set the modification time of a file to a given timestamp
Summary:
Sometimes we want to set the modification time of a parity file to the same as
the source file. For example, distcp copies the parity file without preserving the
modification time, we could easily use this tool to correct the time.

Notice: dfs -touch -m -date '2001-02-03 04:05:06' file is not the right way to
set the modification time of parity file because it's not accurate enough.

Correct way should be:
dfs -stat %Y source_file
]$ 13322800000
dfs -touch -mu '13322800000' parity_file

Test Plan:
add test in TestDFSShell
ant test

Reviewers: tomasz, pritam, hkuang, sdong, avf, dikang

Reviewed By: avf

Task ID: 1165983