[Raid Dir-Raid] Log raided/fixed files in scribe
Summary:
We want to record the files raided or fixed by the raidnode in offline to figure out
affected files if raid bugs are discovered in production.
Because offline encoding and block fixing are done by the mapper tasks, it's better to
log these information to scribe in each mapper. But then we introduce a dependancy
on the scribe and we should closely monitor the availabiliy of scribe logger.

Reviewers: dikang, hkuang

Reviewed By: dikang

Task ID: 1079204