Allow customizing classpath when launching tasks
Summary:
We want the corona task tracker to use apache thrift, but should not
force the tasks to get this classpath onto the tasks launched. This diff
provides a configurable way of removing parts of the task tracker
classpath.

Test Plan:
ran the single node setup with the exclude pattern set and
observed that the thrift jar was getting excluded

Reviewers: aching, pyang

Reviewed By: aching

CC: dms