[Raid Dir-Raid] Raid fsck
Summary:
Raid fsck returns the unrecoverable files under given path.
For directory raid, a file may have multiple stripes, and stripes could be
related to multiple files.
To determine if a corrupt file is recoverable or not, we first locate all related
stripes to the file and collect the number of corrupt blocks for each related
stripe. Then we look at the missing blocks in the file and check if they could be
recoverable. If the stripe contains the missing block has too many corrupt blocks,
we consider the file is unrecoverable.

Test Plan: add a test TestDirectoryRaidShellFsck

Reviewers: hkuang, dikang

Reviewed By: dikang

Task ID: 1079196