Changes to collect a new metrics related to total time taken to fix a file after it is detected to be corrupt.
Summary:
(1) Added a new RaidNodeMetric "numFilesToFixDropped". This stands for the number of files which are detected to be corrupt in the current wake cycle of the corruptionWorker, but that are not submitted for reconstruction since number of pending jobs exceeds the limit.
(2) Logging the "File_Fix_WaitTime" metric. This is the time delay between the detection of a corrupt file and the time when the reconstruction job starts execution.

Test Plan:
For (1) unit test : TestRaidShellFsck_CorruptCounter. The new subtest testRaidNodeMetricsNumFilesToFixDropped() should pass.
For (2) Run TestBlockFixer and observe the mapreduce logs generated in the test output. There will be log entries of the form: RaidMetrics: {"maxPendingJobs":100,"maxFilesPerTask":2,"fileFixWaitTime":6723,"type":"FILE_FIX_WAITTIME"}

Reviewers: dikang, weiyan, hkuang

Reviewed By: weiyan