Enhance the thriftfs API
Summary:
Add new APis to thriftfs so that it allows the ability to create hardLinks, concat, reportBadBlocks, addBlock, etc.

This can also be used to developer a C++ client that can write data to new files in HDFS.

For a quick review, one can look at src/contrib/thriftfs/hadoop.thrift to see the apis that I added.

Test Plan: deploy and run!

Reviewers: ramkumar, rvadali

Reviewed By: rvadali