Fix FastCopy complete() and DAFS create() RPCs
Summary:
FastCopy was using a version of the complete() call which did
not return true when called on an already closed file and instead threw
an exception. This has been fixed by using the right call to complete().

There was another bug where if the client creates a file and the file
create succeeds on the NN, but failover occurs and hence the response
doesn't reach the client. Now the client would retry the same create on
the new Primary after failover and this used to fail with the
AlreadyBeingCreatedException. The handling for this was incorrect in
DAFS and I've corrected it accordingly.

Test Plan:
1) All unit tests.

Revert Plan:

Reviewers: hkuang, sdong, tomasz, weiyan, avf, dikang

Reviewed By: sdong