Implement caching for getPrimaryRegistrationTime() ZK call.
Summary:
During failover each client invokes the
getPrimaryRegistrationTime() call and since this is not cached every
client hits zookeeper. To avoid this problem we have implemented caching
for this call similar to the getPrimaryAddress() call. The caching
behaves as follows :

1) If the cached file does not exist, retrieve result from ZK and
populate it.
2) If this is the first time reading the value, look it up directly from
the cached file.
3) If the cached file exists and this is not the first time we are
reading then if the cache file has changed since the last time we read
it, read the result from the cached file otherwise retrieve the result from ZK.

Test Plan:
1) Unit test performs verification of the number of calls.
2) All avatar tests.

Revert Plan:

Reviewers: tomasz, weiyan, sdong, hkuang

Reviewed By: hkuang

Task ID: 1183425