Remove unnecessary DNS lookups
Summary:
DNS lookups in the status update code path are bad. We hold the global
lock in that path and this can hold up the job tracker.
Also removing the lookups in the initTasks() path, since we do the DNS
lookups now in a separate thread.

Test Plan: unit tests

Reviewers: dms, aching, pyang

Reviewed By: aching