[HDFS] Remove UnsupportedActionException from code
Summary:
UnsupportedActionException is a separate exception that extends IOException
and thus checked. But it is poorly used, and it is never checked for in
try-catch statements. As it doesn't bring in any new logic it makes sense to
use one of general java exceptions instead. IllegalArgumentException seems
to be the best candidate.

Test Plan: ant test

Reviewers: tomasz, hkuang

Task ID: 922523