Fix stuck NameNode caused by a missing edit stream
Summary: This diff fixed NPE caused by a missing edit stream in SyncThread by moving fetching the number of streams closer to where it is used. It fixed the way it handles edit stream error. It also adds a try-catch to catch Throwable that will print the runtime exception and for NameNode to exit when RE happens.

Test Plan: dfstest

Reviewers: tomasz

Reviewed By: tomasz

Task ID: 1236623