Scheduler runtime metric change
Summary:
Currently we only store the runtime of the last run of the
scheduler. The metric makes it hard to set an alarm on when the
scheduler is being slow.
Instead whenever the scheduler is 'running' we should return the time
spent in this run.

Test Plan: Will test that this metric works locally

Reviewers: rvadali, aching

Reviewed By: rvadali

CC: pknowles

Task ID: 1140807