Changes for making perf testing easier
Summary:
This diff lets us make a task tracker usable for profiling tasks.

1. Allow specifying a the command (in place of $JAVA_HOME/bin/java) that
will be used to start the child process. This can be used to start a command
that does the profiling and starts the child. This will allow us to use the
Studio Analyzer collect command for profiling tasks that run on a task tracker.

2. Make corona task tracker get the number of CPUs on the machine through
configuration. This, along with a suitable cpu->resource mapping on the cluster manager
will let us control the number of map slots on the machine.

Test Plan: tested on a single node setup.

Reviewers: dms, pyang

Reviewed By: dms

CC: binu