HDFS: Add a way to disable DFSClientMetrics from being updated
Summary: We realised that the costs of DFSClientMetrics is not small compared to total client costs (sometimes can be 45%) but most of the time, we don't use it (Actually I didn't find where it is used in either data warehouse clusters or HBase). Add an option to disable the metrics to save the costs for cost sensitive applications.

Test Plan: Manually turn it off and make sure normal operations works as normal

Reviewers: hkuang, tomasz, weiyan

Reviewed By: tomasz

CC: dhruba