Enforce validation of include exclude files
Summary:
Added a check for the constructor of HostsFileReader and for refresh
nodes which enforces that the files do exist, if they are configured.
With this check, the refresh nodes would throw an IOException

I'm not sure about the exact logic here... Is it ok that includes and
excludes files are not specified in the conf?

Test Plan:
added test case in TestHostFileReader, but will need to run all test
cases

Reviewers: hkuang

Task ID: 1229592