Adding removeNamespace command
Summary:
The addition of a removeNamespace command:
  removeNamespace nameserviceId [datanodehost:port]

The command will remove the namespace with the given nameserviceId
from the given datanode, or, if the datanode is omitted, from the
datanode running on the local machine.

Test Plan: Will test the command has the expected results (more details to come).

Reviewers: hkuang, sdong

Reviewed By: hkuang

Task ID: 1153125