Reinitialize CM thrift client after returning from waitWhileClusterManagerInSafeMode()
Summary:
When we restart a CM after restoring a state, each job used to get a TException for the first time it tried to connect to the CM after it came out of Safe Mode.
This was because a stale CM thrift client was being used. To solve this problem, in this diff, I am explicitly re-initializing the CM thrift client when we find that the CM is no longer in Safe Mode.

Also, in the getPJTClient() method, I am now creating a client to the PJT only once in the beginning, instead of every time.

Test Plan:
Try to put the CM in Safe Mode, while a job is running. Then bring down the CM and unset the Safe Mode. The job should not get a TException, although it might die
because of other exceptions since we haven't yet committed D519140 and D524393.

Reviewers: dms, rvadali, aching

Reviewed By: dms

Task ID: 1112019