Saving and restoring the NodeManager state
Summary: We are able to save and restore the NodeManager state now.

Test Plan:
I couldn't find a way to test the code completely as such. I tested in two basic ways: (a) Print out what was being read (b) After we recover from the safe mode and
reconstruct the state, persist the state again in a different file, and compare. The second method however requires turning off the compression and turning on pretty
printing.

Reviewers: dms, rvadali, aching

CC: security-diffs@lists

Task ID: 1112019