Fixes double-source problem in hadoop script
Summary:
This is the fix to the following problem.
When starting hadoop instances via `hadoop-daemon.sh` script, it
it at first sources `hadoop-env.sh` file. After some sequence
of actions, it runs `hadoop`, which in turn sources
`hadoop-env.sh` as well. As a result CLI arguments are
doubled.
So far it turned out that some of java arguments mustn't appear
twice in the command line or JVM won't start (particularly I encountered it with -Xrunjdwp
argument). This diff suggests a fix for the problem

Test Plan:
Try to run `./hadoop-daemon.sh` with different `hadoop-env.sh` files: one
with -Xrunjdwp argument inside and one without.

Reviewers: tomasz, hkuang, sdong, avf, mbautin

Reviewed By: tomasz