SessionManager and ClusterManager have the same value of startTime now
Summary:
We tried to solve the addSession() issue in D538438 by persiting the CM's startTime. But SessionManager uses its own value of startTime, hence the problem continued to occur. In this diff, we make sure that both of them have the same value of startTime. Hence, when the CM comes back up, the SessionManager has the original value of startTime. Therefore, any jobs that were assigned a SessionId, but could not make the 'startSession()' request before the restart, won't get an InvalidSessionHandle exception.

Also, when the CM goes into safe mode, if the last assigned SessionId was XYZ.N, the first SessionId assigned after the restart would be XYZ.N+1. Where XYZ is the timestamp of the CM's actual restart, and N is the number of sessions started in the CM's lifetime.

Test Plan:
1. Put the CM in Safe Mode
2. Persist the CM state
3. Restart the CM
4. Ensure that the new SessionIds are as expected.
5. Ensure that a job which has got a sessionId before the CM was in safe mode, and actually gets to submit a startSession() request after the restart, doesn't get an InvalidSessionHandle exception.

Reviewers: dms, rvadali, aching

Reviewed By: dms

Task ID: 1112019