Make Standby.doCheckpoint() private
Summary:
This patch changes the checkpoint handling for the unit test, as the way
it's done now often results in two parallel checkpoints going on - one
that is triggered in the run loop, and one, that is triggered by the
unit test. I changed this to be handled by injection handling.

Test Plan: running two affected tests

Reviewers: hkuang, pritam

Reviewed By: pritam