Fix race condition in Avatar refresh namenodes.
Summary:
There was a race condition in the refreshNameNodes method of
AvatarDataNode where refreshNamenodes() would stop the servicePair
thread to a particular namenode and would not wait for that thread to
exit. Now when this thread exited it would remove out a particular
namenode from the namenodeThreads map. The problem with this is that
before the thread could exit another refreshNode call might add
that namenode in and when our old servicePair thread exits it removes
the namenode from the list of threads causing an inconsistency. Hence
I've added a join() on the thread for this purpose and also put it under
the refreshNamenodesLock() to avoid any race conditions.

Test Plan:
Run TestAvatarRefreshNamenodes 10 times which was failing because of
this race condition.

Reviewers: hkuang, weiyan, sdong, tomasz

Reviewed By: tomasz

Task ID: 1338455