Fix some issues with hftp
Summary:
This patch introduces a few fixes to hftp:

1) Add metrics for the number of http requests for
*) list files
*) data (open a file)

the second one was the reason of yesterday slowdown

2) fix bug with initialization of http server which resulted in the hftp
requests go to the avatar port (9001) -> to which all incremental block
reports, and block reports, and heartbeats are sent by the datanodes.

With a high number of hftp requests, the datanode port was imply
trashed, and as a result we had 500 threads serving client requests.
Now the hftp requests go to the client port (9000)

3) For the listPaths and data requests, I introduced a cached rpc
handle, which should significantly reduce overhead of the hftp calls (it
at least saves one rpc for each request + the time to setup the proxy)

Test Plan: manually on dev cluster

Reviewers: hkuang

Reviewed By: hkuang

Task ID: 1353359