ConfigManager to fail init if there's no file
Summary:
We should have a distinction between the initialization of the
config manager and the consecutive reloads of the config.
If the config is missing during the initialization we should fail the
startup.
If the config goes missing while the cluster manager is running it is
not that big of a deal.

Test Plan: unittests

Reviewers: aching, rvadali

Reviewed By: aching