Fix the NPE in ClusterManager
Summary:
Peter hit this exception when starting up the corona cluster.
If we can't generate the config during the startup sequence it seems
that the pools file is null, then the startup fails.
At the same time it is OK to just return false on this one since it will
get reloaded in the future.

Test Plan: unittests

Reviewers: rvadali, aching

Reviewed By: rvadali

CC: pknowles