Decouple failover logic from DistributedAvatarFileSystem.
Summary:
For protocols other than ClientProtocol, we need to use the
same failover logic. For example for the balancer we might need the same
piece of logic for NamenodeProtocol to do a failover. For this purpose
I'm decoupling the failover logic so that it can be reused easily.

I also removed some zookeeper watcher stuff that was not being used.

Test Plan:
All avatar unit tests.

Revert Plan:

Tags:

- begin *PUBLIC* platform impact section -
Bugzilla: #
- end platform impact -

Reviewers: hkuang, weiyan, sdong, tomasz

Reviewed By: hkuang

Task ID: 1142725