Kill the JVM of the task reported done
Summary:
This fixes the problem of the JVMs that are running even after
the task itself is reported is done.
Specifically it was happening because there were threads that are
non-daemon and they keep on running even after the main() part of the
taks finishes.

This way we create a 'kill' command to be run X seconds later and kill
the JVM is the task is reported as done. We were already doing that but
only when the jvm shell execution exited, this also does it when the
java process is still running and not returning, but the task is
reported as done.

Test Plan:
I tested it manually in my single node setup. Having a job
that simulates the problem (launching threads that won't die after the
map task is done).

The Child processes were successfully destroyed.

Reviewers: aching, rvadali

Reviewed By: aching

Task ID: 1349122