Namespace Notifier initial version
Summary:
Initial version of the namespace notifier, with most of the
functionality in place: reading from the edit log,
notifying one or multiple clients, detection of a failed
client by the server, detection of a failed server by a client
(the heartbeat system), saving notifications in the server history,
reconnecting by the client when a server comes back up.

Test Plan: Manual testing + unit tests for all the server side components.

Reviewers: sdong, hkuang, tomasz, jaxlaw

Reviewed By: sdong