Allow corona to run without pools file
Summary:
This handles the case when corona is configured with json and
no pools file.
We do need that sometimes since we are setting up one-off clusters with
no need for pools, but the corona cluster fails to start now.

Test Plan: Peter is setting up a cluster now that triggered the issue.

Reviewers: aching, rvadali, pyang

Reviewed By: aching

CC: pknowles