Make standalone JT configurable
Summary:
Right now the configuration that is used to launch a remote
jobtracker is set cluster-wide. This is undesirable since we have large
jobs that need more memory and potentially other command line things
that we want to configure per job.

This patch makes it so that the configuration of the job is applied to
the configuration object used when constructing the
CoronaJobTrackerRunner

Test Plan: Just tested in the single node hadoop

Reviewers: rvadali, aching

Reviewed By: rvadali

Task ID: 1405185