Only kill the JVM if it ran all the tasks
Summary:
Since the JVMs get stuck sometimes after running tasks we
decided to start killing them after the task is finished. However it
gets in the way of the JVM reuse logic.
This is failing the tests.

Test Plan:
ant test (actually ant test succeeds on my box anyway, but it
fails in the automated system)

Reviewers: aching, rvadali

Reviewed By: rvadali