Exception caused writeLock to be held indefinitely.
Summary:
In processReport and processBlocksBeingWritten() we have a
finally block that does checkSafeMode() and then releases the writeLock.
Now if checkSafeMode throws a RunTimeException then this would cause the
writeLock to never be unlocked and no thread would ever be able to
acquire the writeLock.

Test Plan: All unit tests.

Reviewers: hkuang, tomasz

Reviewed By: hkuang

Task ID: 1404895