Fix wrong standby behaviour at startup
Summary:
After debugging problems with darklaunch with Pritam, we discovered a
very bad behaviour of standby on startup, immediately follwoed by
failover.

1) When Standby starts, it creates an empty edits file
2) recover transition read, does not save the namespace, but rolls the
log, so the standby writes to edits.new
3) when the FIRST checkpoint starts, it closes the log (edits.new) and starts
saving namespace
4) When failover is triggered, the save NS is cancelled, and the edit
log is reopened -> the edit log ALWAYS reopens to edits (!!!)
5) The failover completes and the standby - now the new primary - writes to
edits, which is newer than edits.new (!!!)

This patch reverts the previous behaviour on startup with empty edits.
In this situation we do not roll the log, but just append to edits.

This bug happens only when failover is triggered after
Standby.doCheckpoint closes the log, and before is finishes to save the
namespace ONLY during the FIRST checkpoint after standby's startup.

Test Plan:
adjusted revious test cases to ensure that the startup with empty edits
appends to edits, and does not roll the log

I will try to add a testcase when the failover happens in the situation
described above, but this "startup rolling" logic will be gone anyways
once I start merging the new layout. There this problem does not exist.

running all test cases now

Reviewers: hkuang, pritam

Reviewed By: pritam