Enable checkpointing for TestStandbySafeMode
Summary:
This patch enables checkpoint for the sait testcase.

I think it's a very bad practice that checkpointing is disabled in some
test cases, as it does not represent the real scenario. In this case, my
earlier patch caused that in one scenario, the standby was not able to
consume edits.new because of disabled checkpointing.

Test Plan: ant test TestStandbySafeMode

Reviewers: hkuang, pritam

Reviewed By: pritam

Task ID: 1417796