[hdfs] Add file creation time meta data
Summary:
The changes include completeFileInternal, setTimesInternal, startFileInternal. In those functions, accessTime change is disabled. setTimes, append/overwrite files, and close will not change the accessTime.

For touch, option -a and option -d have no impact to accessTime anymore. If a file is created by -a or -d, the file's accessTime is the time system creates the file. setTimes after that call has no impact.

For copyFiles, the dest files' accessTime is the time that the system creates the file. setTimes to the srcFile's accessTime is ignored.

Test Plan: A new test class TestAccessTime is added. And the following test cases have been updated: TestCopyFiles.java, TestDFSShell.java, TestSetTimes.java

Reviewers: sdong, hkuang, weiyan

Reviewed By: sdong

Task ID: 1131005