fix synchronization in ipc.Server
Summary:
There's a bug in closing ipc.Server where we are not accessing
a list in a thread-safe manner. This causes and ArrayOutOfBounds, I
don't think it happens often since we haven't seen this in the wild
really, but there was one report that helped me identify the bug.

Test Plan: Just running unittests should be enough in this case

Reviewers: aching, hkuang, rvadali

Reviewed By: hkuang

CC: mattwkelly

Task ID: 1321354