Reducers should fail map tasks faster when the error is serious
Summary:
The mapper task tracker now reports whether it could find the map output or not when servicing the http request.  This gives the reducer a strong signal of needing to rerun the task.

The reducer task is now aware of whether the map output was found or not.  If it isn't found, it will notify the jobtracker immediately.

Test Plan:
Passed all corona unittests.

Need to try out on the mrtest cluster.

Reviewers: dms, rvadali, kevinwilfong

Reviewed By: rvadali

Task ID: 1426074