Fix TestUnderreplicatedBlocks
Summary:
This patch extends sleep during the test.

I think there was a race here, where an excess replica was scheduled for
deletion late, then the IBR would not
contain the info about this block, and finally the block could not be
replicated back to the datanode (this is because of the delete IBR retry
which is quite long).

Test Plan: ant test-core

Reviewers: hkuang, pritam

Reviewed By: pritam