[L38-5] Introduce MD5 utilities
Summary:
With the L38 layout, the md5 of each image is stored in a separate
file.
We can no longer store md5 in version file, since we can have
multiple
images in a storage directory.

This patch introduces utilities for writing a computed md5 into
a file.

AtomicFileInputStream is a class to guarantee that the hash
is written
atomically (using temporary file, and rename).

MD5FileUtils is a class for writing, reading, and
validating md5
digests.

Test Plan: added two test suites

Reviewers: hkuang, avf

Reviewed By: avf

Differential Revision:

Task ID: 1009397