[L38-11-elopt] Port edit log optimizations.
Summary:
With this patch, we are going to have substantial savings on writing
to
edit log, as all longs, shorts, will be written in binary form
without
expensive transformations into utf8 strings (coding,decoding). Also
for
each operations, we will not be writing the length of writables.
This
change is compatible with Apache image layout -39 -> in our case it
will
be -40, since we have a shift for hardlinks.

Test Plan: will run ant-test-core

Reviewers: hkuang, avf

Reviewed By: avf

Differential Revision:

Task ID: 883594

Summary:

Test Plan:

Reviewers:

CC:

Task ID: #

Blame Rev: