[L38-12-bumpavatar] Plugin journal manager into avatarnodes
Summary: This patch changes the standby mechanism of ingesting edits,
and plugs in the journal manager interface.With this patch we have a
full flavour implementation of journal manager, the standby is doing
checkpoints, and also is able to failover.

****I added mechanism for recovering from ingest failure, after several
failures, the state of stanby is reset, and we attempt to recover from
problems

****With this patch, it is possible to save namespace when standby is
up, and standby should proceed in this situation, so there is no need
for handling stale checkpoints etc.

----------------------------------------------

Problems:

1) What should happen when we are quiescing standby to -1 transaction
id. Since I try to obtain the streams gradually with retries, the last
stream that does not exist will cause 30 retries before the standby
quiesces

2) The biggest problem that I see now is with the catchUp concept. For
now it's implemented in a way where ingest looks if it's consuming the
inprogress segement. If so, it checks the size of the stream and the
current position and compares them. If it's consuming a finalized
segment, we assume that it's caching up. This means that BK
EditLogInputStream must implement lenght() method for in progress
segments at least, which would give byte size of the stream.

----------------------

All tests, including new ones work, but I'm still working on this patch.
Added tests for
-ingest recover
-reprocessing ingest
-retries when obtaining input stream

--------------------

I removed snapshot from compiling, because the new layout is not there
at all!! I also made some changes in OfferService.java and Client.java,
which are in the trunk, which fixed the timing out test cases.

Test Plan: ant test, manually new test cases

Reviewers: avf, hkuang

Reviewed By: avf

    Task ID: 883594