HDFS Fsck doesn't show total number of data nodes or racks if user asks for file level of information or further
Summary:
Now Fsck always prints out total number of data nodes and racks. It is a very expensive operation.
Change the codes not to print it out if user asks for file or block level information. It's very unlikely that user are interested in total number of nodes when having "-files" specified.

Test Plan: manual test

Reviewers: hkuang, tomasz, weiyan

Reviewed By: weiyan