HDFS AvatarDataNode to resolve DNS address on some failures
Summary: We've made then change in RPC client so that if DNS changes, it will resolve it and use the new address. However, when the proxy is not yet successfully created, this will not be triggered so data node will stuck there. This change adds a resolve logic too.

Test Plan: Run "ant test" to make sure it doesn't break anything.

Reviewers: tomasz, pritam, hkuang

Reviewed By: pritam