Shutdown responder cleanly on avatarnode shutdown
Summary:
This patch changes the behaviour of shutting down rpc server.

Previously
1) running = false causes the connections to be cleaned and sockets to
be closed
2) handlers can still make changes to the namespace
3) responses cannot be sent

Now

1) listener is shutdown but it does not cleanup connections
2) reponder is put into extension mode
3) running = false
4) handlers can finish
5) responses are sent to clients

------------------

This will help with tests, but also should save some retry logic on the
clients during failover in which the clients don't know if a call
succeeded.

------------------

Also, increased heap space for test, as with large output the tests
crash.

Test Plan:
added TestAvatarShutdown in which a number fo files is created, but the
handlers hold untill a shutdown command comes. Then they proceed, and
the repsonses are sent to clients.

After, we restrt the cluster and all the files must be present in the
namespace

Reviewers: hkuang, pritam

Reviewed By: pritam